Final NPC_dialogue change (Branch COMPLETE)
Finally implemented the thing where the player can't move if the dialogue UI is on screen. The problem is, the way stopping the player for certain things is implemented, the movement physics don't work anymore. The player just moves without any physics, so I gotta fix that. But I'll do it in the movement_&_polish branch.