Created the basic assets
Just created the basic assets. I gotta understand when, where, and how these papers appear in order to create a structure for other objects to call/create it and for it to exist as an interactable on its own. It's confusing, so I gotta play the game again and figure it out