Mechanic implementation COMPLETE
I still have to set up the rest of the opacity lerps (A few of which I will only be able to set up after Kimya finishes the paper ripping animation), and I still have to make it so that the item managers stop detecting input when these UI elements are visible.

But other than that, we gucci.